v0.97 - beta 1 - Добавлен Bluetooth протокол работы устройств - Реализована схема работы устройств, питающихся от батареек, с беспроводными интерфейсами - Наследование сенсоров переделано так, чтобы исключить повторную инициализацию каждого из них. Для этого инициализация сенсора перенесена в отдельный метод Init. Вызов метода контроллируется через базовый класс Sensor. - Добавлен класс Wireless, являющийся абстракцией для всех беспроводных интерфейсов - Увеличение скорости подключения к WiFi за счет кеширования данных точки доступа. - Дополнительная очередь UDP broadcast пакетов в WebServer - Отказ от Bluetooth (отключен в menuconfig, может быть включен там же) - Настройки команд и сенсоров сгрупированы и перенесены в Settings.cpp сенсоры и команды - Изменен формат хранения данных о GPIO для сенсоров и команд в устройстве

+++ libs/DateTime/DateTime.h
@@ -37,7 +37,7 @@ class Time {
     static int8_t     TimezoneOffset;
 
     static uint32_t   Uptime();
-    static uint32_t   UptimeU();
+    static uint64_t   UptimeU();
     static uint32_t   Unixtime();
     static DateTime_t DateTime();
     static string     UnixtimeString();
